+++ lib/page/user/bookmark/bookmark_page.dart
@@ -16,15 +16,14 @@
 
 import 'dart:async';
 
-import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:pixez/component/sort_group.dart';
-import 'package:pixez/i18n.dart';
+import 'package:pixez/constants.dart';
 import 'package:pixez/lighting/lighting_page.dart';
 import 'package:pixez/lighting/lighting_store.dart';
 import 'package:pixez/main.dart';
 import 'package:pixez/network/api_client.dart';
-import 'package:pixez/page/user/bookmark/tag/user_bookmark_tag_page.dart';
+import 'package:pixez/page/user/bookmark/fluent_state.dart';
+import 'package:pixez/page/user/bookmark/material_state.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class BookmarkPage extends StatefulWidget {
@@ -40,18 +39,23 @@ class BookmarkPage extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  _BookmarkPageState createState() => _BookmarkPageState();
+  BookmarkPageStateBase createState() {
+    if (Constants.isFluentUI)
+      return FluentBookmarkPageState();
+    else
+      return MaterialBookmarkPageState();
+  }
 }
 
-class _BookmarkPageState extends State<BookmarkPage> {
+abstract class BookmarkPageStateBase extends State<BookmarkPage> {
   late LightSource futureGet;
   String restrict = 'public';
-  late RefreshController _refreshController;
+  late RefreshController refreshController;
   late StreamSubscription<String> subscription;
 
   @override
   void initState() {
-    _refreshController = RefreshController();
+    refreshController = RefreshController();
     restrict = widget.restrict;
     futureGet = ApiForceSource(
         futureGet: (e) =>
@@ -59,7 +63,7 @@ class _BookmarkPageState extends State<BookmarkPage> {
     super.initState();
     subscription = topStore.topStream.listen((event) {
       if (event == "302") {
-        _refreshController.position?.jumpTo(0);
+        refreshController.position?.jumpTo(0);
       }
     });
   }
@@ -67,7 +71,7 @@ class _BookmarkPageState extends State<BookmarkPage> {
   @override
   void dispose() {
     subscription.cancel();
-    _refreshController.dispose();
+    refreshController.dispose();
     super.dispose();
   }
 
@@ -79,7 +83,7 @@ class _BookmarkPageState extends State<BookmarkPage> {
           children: [
             LightingList(
               source: futureGet,
-              refreshController: _refreshController,
+              refreshController: refreshController,
               header: Container(
                 height: 45,
               ),
@@ -97,56 +101,5 @@ class _BookmarkPageState extends State<BookmarkPage> {
     }
   }
 
-  Widget buildTopChip(BuildContext context) {
-    return Align(
-      alignment: Alignment.topCenter,
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          SortGroup(
-            children: [I18n.of(context).public, I18n.of(context).private],
-            onChange: (index) {
-              if (index == 0)
-                setState(() {
-                  futureGet = ApiForceSource(
-                      futureGet: (bool e) => apiClient.getBookmarksIllust(
-                          widget.id, restrict = 'public', null));
-                });
-              if (index == 1)
-                setState(() {
-                  futureGet = ApiForceSource(
-                      futureGet: (bool e) => apiClient.getBookmarksIllust(
-                          widget.id, restrict = 'private', null));
-                });
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 8.0),
-            child: InkWell(
-              onTap: () async {
-                final result = await Navigator.of(context).push(
-                    MaterialPageRoute(builder: (_) => UserBookmarkTagPage()));
-                if (result != null) {
-                  String? tag = result['tag'];
-                  String restrict = result['restrict'];
-                  setState(() {
-                    futureGet = ApiForceSource(
-                        futureGet: (bool e) => apiClient.getBookmarksIllust(
-                            widget.id, restrict, tag));
-                  });
-                }
-              },
-              child: Chip(
-                label: Icon(Icons.sort),
-                backgroundColor: Theme.of(context).cardColor,
-                elevation: 4.0,
-                padding: EdgeInsets.all(0.0),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+  Widget buildTopChip(BuildContext context);
 }
